+++ epyk/tests/data_urls.py
@@ -79,6 +79,8 @@ COVID_US = "https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-co
 COVIS_UK = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/country_data/United%20Kingdom.csv"
 COVIS_CTY = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/excess_mortality/excess_mortality.csv"
 
+OWID_BIRTH_OUT_MARRIAGE = "https://raw.githubusercontent.com/owid/owid-datasets/master/datasets/Births%20outside%20of%20marriage/Births%20outside%20of%20marriage.csv"
+
 HEART_FAILURE_RECORDS = "https://raw.githubusercontent.com/rsharankumar/Learn_Data_Science_in_100Days/master/Data/heart_failure_clinical_records_dataset.csv"
 COUNTRY_WISE_COVID = "https://raw.githubusercontent.com/rsharankumar/Learn_Data_Science_in_100Days/master/Data/country_wise_covid.csv"
 NETFLIX_TITLE = "https://raw.githubusercontent.com/rsharankumar/Learn_Data_Science_in_100Days/master/Data/netflix_titles.csv"